Added OS and docker configuration classes
